refactor: move all core location APIs to location package
